Add features, driver. Refactor
Create a full model in place of the Game,
Emulator association table to allow features
and a driver to be added.

Reorder when things are added and committed to
avoid warnings when the database is created.
Commit regularly rather than following the
processing of each DAT file to speed up
performance.

Create a general purpose function for getting
existing records by attribute and use in all
cases.