Process MAME DATs and handle cross-references
Re-order some of the session.add() and session.commit() calls
to avoid warnings about objects not being added.

Fix process_games so it adds the correct game id to the
references dicts. Handle all but a handful of game->game
references (some investigation required into the remaining
unhandled cases).

Don't try to handle applying references in the same step as
adding games. This is a separate step at the end. Accept
it's a bit slower in exchange for reduced complexity.

Remove the function that tried to handle references on the
fly and remove associated tests.