Refactor rom retrieval/creation
Build runs up to this point showed multiple roms being returned with
same name, size and crc. This was because some rom elements do not
have a crc. Matches were therefore made only on name and size.

This commit refactors the rom retrieval and creation to use the
same approach as for disks.