Expunge session after processing each game
Breaking up the GitHub action into separate jobs and expunging the
session after each DAT has been processed delayed the workflow
failure but didn't eliminate it (happens at around DAT 160 vs 147).

This commit should avoid a situation where a single DAT is so large
that it's enough to cause the runner to run out of memory. It
seems to be a bit slower.