Add game references in main loop
Replace functions which added references via references
dictionary at the end with addition during processing of
each game.

It's not impossible it may be needed to catch unhandled
references and add at the end in the future if the rules
regarding relationships between games change in future
DATs, or there is some level of complexity which can't
be accomodated in the clone sorting script. In which case,
the code in the commit(s) before this one can be
re-implemented (in addition to the new approach, which
is preferable wherever possible).