Add disk records to GameEmulator not Game
Disks are now added to the join table between Game and Emulator,
rather than Game itself. This means the main index on Game includes
only rom signatures, making it easy to query using a zipfile as the
source. It also prevents multiple Game records being created just
because the disk (which is in a different CHD file) happens to be
different.