Build up data with pandas
By using a meaningful index for most tables, we can rely on pandas drop
duplicates method to deal with duplicate games/roms etc which is much
faster than querying a database and adding relationships.

The new functions for building tables are not tested and have not yet
been written to a database. This is a first prototype.